compling http_client sample ok on Linux 32 bits

+++ lib_acl_cpp/samples/aio/http_client/main.cpp
@@ -141,7 +141,7 @@ protected:
 		if (!compressed_ || this->is_unzip_body()) {
 			(void) write(1, data, dlen);
 		} else {
-			printf(">>>read body: %ld\r\n", dlen);
+			printf(">>>read body: %ld\r\n", (long) dlen);
 		}
 		return true;
 	}
